requirements.txt
asgiref==3.3.1
autopep8==1.5.4
Django==3.1.4
Pillow==8.0.1
pkg-resources==0.0.0
pycodestyle==2.6.0
pytz==2020.4
qrcode==6.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
wifi-qrcode-generator==0.1
